Enhances reserve button functionality
Improves the reserve button by making it redirect to a configurable call-to-action URL.

The URL is retrieved from the Syndified plugin settings, allowing dealers to customize the destination for each product. It also includes product name and article ID as parameters, mirroring the Syndified templates.

If no URL is configured, it falls back to the product permalink.
The color of reserve button is now customizable from dealer settings.

Also, this commit removes unnecessary translation from the reserve icon.